refactor(enter_game): OpenGame 吸收 wait+HDR 成完整 op,OpenAndEnterGame 简化为 2 节点
